first

+++ src/kaocommon/kaofield.go
@@ -34,7 +34,7 @@ type AtReqColumn struct {
 	Wide    		string	`db:"wide"`
 	Send_group    	*string	`db:"send_group"`
 	Supplement    	string	`db:"supplement"`
-	Price           *int64	`db:"price"`
+	Price           *int	`db:"price"`
 	Currency_type   string	`db:"currency_type"`
 	Title           string  `db:"title"`
 }
